feat: full IC2 Port implementation LV-EV (v0.1.0-SNAPSHOT)
Ship complete mod from skeleton to endgame: EU network, 20+ machines,
nuclear/fusion reactors, 18+ crops, nano/quantum armor with modules,
fluid cells, ESU, miner/pump, teleporter/terraformer/replicator.

Add professional README, CHANGELOG, GitHub Actions CI, advancements,
240+ recipes, EN/RU localization, and machine development template.

+++ src/main/java/dev/ic2port/api/energy/IEnergyNode.java
@@ -16,12 +16,12 @@ public interface IEnergyNode {
     /**
      * @return maximum amount of EU this node can store or buffer
      */
-    int getCapacity();
+    double getCapacity();
 
     /**
      * @return current amount of EU stored in this node
      */
-    int getStoredEnergy();
+    double getStoredEnergy();
 
     /**
      * @return voltage tier of this node (1 = LV, 2 = MV, 3 = HV, 4 = EV)
